// src/api/generated/model/productAdminResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */
import type { CategoryResponse } from "./categoryResponse.ts";
import type { OptionRequest } from "./optionRequest.ts";
import type { ProductVariantResponse } from "./productVariantResponse.ts";

export interface ProductAdminResponse {
	id?: string;
	name?: string;
	category?: CategoryResponse;
	options?: OptionRequest[];
	variants?: ProductVariantResponse[];
}